+++ lib/src/main/kotlin/tkhamez/discordRelay/lib/Config.kt
@@ -11,9 +11,9 @@ object Config {
 
     var isBotToken = false
 
-    var channelIds = ""
+    var channelIds = mutableListOf<String>()
 
-    var onlyMentionEveryone = false
+    var onlyMentionEveryone = mutableListOf<Boolean>()
 
     var webhook = ""
 
@@ -27,9 +27,44 @@ object Config {
      */
     var gatewayResponseBot: MutableMap<String, HttpResponseGateway> = mutableMapOf()
 
-    val channels = mutableListOf<GuildChannel>()
+    /**
+     * Cache of channel IDs and names with guild.
+     */
+    val guildChannels = mutableListOf<GuildObject>()
 
-    fun channelIdList(): List<String> {
-        return channelIds.split(',')
+    /**
+     * Cache of role IDs and names with guild.
+     */
+    val guildRoles = mutableListOf<GuildObject>()
+
+    /**
+     * Make sure that the onlyMentionEveryone list has the same size as the channelIds list.
+     */
+    fun sanitize() {
+        if (channelIds.size > onlyMentionEveryone.size) {
+            repeat(channelIds.size - onlyMentionEveryone.size) {
+                onlyMentionEveryone.add(true)
+            }
+        } else if (onlyMentionEveryone.size > channelIds.size) {
+            repeat(onlyMentionEveryone.size - channelIds.size) {
+                onlyMentionEveryone.removeLast()
+            }
+        }
+    }
+
+    /**
+     * Return list of channel IDs without additional text at the end.
+     */
+    fun channelIds(): MutableList<String> {
+        val ids = mutableListOf<String>()
+        for (value in channelIds) {
+            val pos = value.indexOf(' ')
+            if (pos > 0) {
+                ids.add(value.substring(0, pos))
+            } else {
+                ids.add(value)
+            }
+        }
+        return ids
     }
 }